Add support for Debian and Kali Linux AMIs
This commit adds support for Debian and Kali Linux AMIs in the `GetLatestAMI` function. The corresponding owner IDs for these distributions have been included in the `distToOwner` map.

Additionally, an example function `ExampleIsEC2Instance` has been added to demonstrate the usage of the `IsEC2Instance` function. This function checks if the code is running on an EC2 instance and logs the appropriate message.

Some test functions have been commented out as they are not currently needed.