Add support for asdf and update labels
This commit introduces support for asdf, a tool for managing multiple runtime versions on a per-project basis. The added code sets up asdf for Go, Ruby, and Python used by the project. It also defines the URLs and paths for setup_asdf.sh and sources it.

In addition, the commit updates the labels.yaml file in the .github directory. New labels are added for various areas of the project, including dynamo, secretsmanager, ssm, ec2, s3, magefiles, go-mod-sum, and asdf. These labels provide better categorization and organization of changes made to different parts of the project.

Some other changes related to renovate and gvm have been made, but they are not mentioned in the commit message.